refactor: Move FakeMoviesRepository to core/data module
This commit moves the `FakeMoviesRepository` from the `features/movies` module to the `core/data` module.

This refactoring makes the fake implementation available for use in other feature modules, promoting code reuse and centralizing test utilities within the data layer. The `MoviesViewModelTest` has been updated to import the class from its new location.
