+++ features/movies/src/commonTest/kotlin/MoviesViewModelTest.kt
@@ -1,19 +1,15 @@
 package com.elna.moviedb.feature.movies
 
-import com.elna.moviedb.core.data.movies.MoviesRepository
+import com.elna.moviedb.core.data.movies.FakeMoviesRepository
 import com.elna.moviedb.core.model.AppResult
 import com.elna.moviedb.core.model.Movie
 import com.elna.moviedb.core.model.MovieCategory
-import com.elna.moviedb.core.model.MovieDetails
 import com.elna.moviedb.feature.movies.model.MoviesEvent
 import com.elna.moviedb.feature.movies.model.MoviesUiAction
 import com.elna.moviedb.feature.movies.model.MoviesUiState
 import com.elna.moviedb.feature.movies.ui.movies.MoviesViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.advanceUntilIdle
@@ -372,53 +368,4 @@ class MoviesViewModelTest {
         title = title,
         posterPath = "/path$id.jpg"
     )
-}
-
-/**
- * Fake implementation of MoviesRepository for testing.
- * Provides controllable behavior for all repository operations.
- */
-private class FakeMoviesRepository : MoviesRepository {
-    private val moviesFlows = mutableMapOf<MovieCategory, MutableStateFlow<List<Movie>>>()
-    private val nextPageResults = mutableMapOf<MovieCategory, AppResult<Unit>>()
-
-    val loadNextPageCallCount = mutableMapOf<MovieCategory, Int>()
-    var clearAndReloadCallCount = 0
-    var clearAndReloadDelay = 0L
-
-    init {
-        // Initialize flows and counters for all categories
-        MovieCategory.entries.forEach { category ->
-            moviesFlows[category] = MutableStateFlow(emptyList())
-            loadNextPageCallCount[category] = 0
-        }
-    }
-
-    fun setMoviesForCategory(category: MovieCategory, movies: List<Movie>) {
-        moviesFlows[category]?.value = movies
-    }
-
-    fun setNextPageResult(category: MovieCategory, result: AppResult<Unit>) {
-        nextPageResults[category] = result
-    }
-
-    override suspend fun observeMovies(category: MovieCategory): Flow<List<Movie>> {
-        return moviesFlows[category] ?: flowOf(emptyList())
-    }
-
-    override suspend fun loadMoviesNextPage(category: MovieCategory): AppResult<Unit> {
-        loadNextPageCallCount[category] = (loadNextPageCallCount[category] ?: 0) + 1
-        return nextPageResults[category] ?: AppResult.Success(Unit)
-    }
-
-    override suspend fun getMovieDetails(movieId: Int): AppResult<MovieDetails> {
-        return AppResult.Error("Not implemented in fake")
-    }
-
-    override suspend fun clearAndReload() {
-        clearAndReloadCallCount++
-        if (clearAndReloadDelay > 0) {
-            kotlinx.coroutines.delay(clearAndReloadDelay)
-        }
-    }
 }
